requirements/base.txt
# Base
pytz==2019.2
python-slugify==3.0.3
Pillow==6.1.0
psycopg2==2.8.3 --no-binary psycopg2

# Django
django>=2.2.3

# Django REST Framework
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.1

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==19.1.0

# Static files
whitenoise==4.1.3

# Celery
redis>=3.3.0
django-redis==4.10.0
celery==4.2.1
flower==0.9.3
tornado>=4.2.0,<6.0.0

# JWT
pyjwt==1.7.1

# Tools
django-extensions==2.2.1
